5565: Added links

+++ app/helpers/meals_helper.rb
@@ -29,6 +29,8 @@ module MealsHelper
       a << :close if meal.closeable?
       a << :finalize if meal.finalizable?
       a << :destroy
+      a << :contact_diners
+      a << :contact_team
       options[:except].each{ |x| a.delete(x) }
     end
 
@@ -53,6 +55,10 @@ module MealsHelper
       when :destroy
         links << link_to(icon_tag("trash") << name, meal_path(meal), title: title, method: :delete,
           data: { confirm: I18n.t("activerecord.delete_confirms.meal", title: title) })
+      when :contact_diners
+        links << link_to(icon_tag("envelope") << name, new_meal_message_path(meal, r: "diners"), title: title)
+      when :contact_team
+        links << link_to(icon_tag("envelope") << name, new_meal_message_path(meal, r: "team"), title: title)
       end
     end
